Add files via upload
The file __init__.cpython-313.pyc is a compiled Python file that belongs to the internal structure of the pip package manager, specifically within its importlib metadata module. This module plays a key role in how pip retrieves and manages metadata for installed Python packages using Python’s importlib system. It includes definitions related to environments and distributions, which help pip identify available packages and their attributes within a given Python environment. As a .pyc file, it contains bytecode that the Python interpreter uses to execute the module efficiently, typically generated automatically when a Python script is run. This file is part of the virtual environment setup and contributes to pip's ability to function smoothly without needing to recompile source files every time.